Remove CallBaseToken and add some docs
