+++ src/main/java/kroppeb/stareval/parser/Parser.java
@@ -4,7 +4,6 @@ package kroppeb.stareval.parser;
 import kroppeb.stareval.token.AccessToken;
 import kroppeb.stareval.token.AccessableToken;
 import kroppeb.stareval.token.BinaryOperatorToken;
-import kroppeb.stareval.token.CallBaseToken;
 import kroppeb.stareval.token.ExpressionToken;
 import kroppeb.stareval.token.IdToken;
 import kroppeb.stareval.token.NumberToken;
@@ -15,9 +14,7 @@ import org.jetbrains.annotations.NotNull;
 
 import net.minecraft.client.util.CharPredicate;
 
-/**
- *
- */
+
 public class Parser {
 	final ParserOptions options;
 	final ParserOptions.ParserParts parserParts;
@@ -32,8 +29,8 @@ public class Parser {
 	}
 	
 	ExpressionToken parseInternal(final StringReader input) throws Exception {
-		// parser state
-		final ParserState state = new ParserState();
+		// parser stack
+		final TokenStack stack = new TokenStack();
 		
 		
 		while (input.canRead()) {
@@ -41,30 +38,30 @@ public class Parser {
 			
 			if (parserParts.isIdStart(c)) {
 				Token token = parseIdGroup(input);
-				state.push(token);
+				stack.push(token);
 			} else if (parserParts.isNumberStart(c)) {
 				// start parsing a number
 				final String numberString = readWhile(input, parserParts::isNumberPart);
-				state.push(new NumberToken(numberString));
+				stack.push(new NumberToken(numberString));
 			} else if (c == '(') {
-				state.push(new UnfinishedArgsToken());
+				stack.push(new UnfinishedArgsToken());
 			} else if (c == ',') {
-				state.commaReduce();
+				stack.commaReduce();
 			} else if (c == ')') {
-				state.bracketReduce();
+				stack.bracketReduce();
 			} else {
-				if (state.peek() instanceof ExpressionToken) {
+				if (stack.peek() instanceof ExpressionToken) {
 					// maybe binary operator
 					OpResolver<BinaryOp> resolver = options.binaryOpResolvers.get(c);
 					if (resolver != null) {
-						state.push(new BinaryOperatorToken(resolver.check(input)));
+						stack.push(new BinaryOperatorToken(resolver.check(input)));
 						continue;
 					}
 				} else {
 					// maybe unary operator
 					OpResolver<UnaryOp> resolver = options.unaryOpResolvers.get(c);
 					if (resolver != null) {
-						state.push(new UnaryOperatorToken(resolver.check(input)));
+						stack.push(new UnaryOperatorToken(resolver.check(input)));
 						continue;
 					}
 				}
@@ -72,9 +69,9 @@ public class Parser {
 				throw new Exception("unknown char: '" + c + "'");
 			}
 		}
-		ExpressionToken result = state.expressionReducePop();
-		if (!state.stack.isEmpty()) {
-			throw new Exception("stack isn't empty: " + state.stack + " top: " + result);
+		ExpressionToken result = stack.expressionReducePop();
+		if (!stack.stack.isEmpty()) {
+			throw new Exception("stack isn't empty: " + stack.stack + " top: " + result);
 		}
 		
 		return result;
@@ -83,19 +80,14 @@ public class Parser {
 	@NotNull
 	private Token parseIdGroup(StringReader input) throws Exception {
 		final String id = readWhile(input, this.parserParts::isIdPart);
-		if(!input.canRead())
-			return new IdToken(id);
-		
-		char c = input.peek();
-		
-		if (c == '(') {
-			// TODO do we really need to make this a separate token?
-			//		I think I did it cause something with the access
-			return new CallBaseToken(id);
-		}
-		
 		AccessableToken token = new IdToken(id);
-		
+
+		if(!input.canRead()) {
+			return token;
+		}
+
+		char c = input.peek();
+
 		if (c == '.') {
 			do {
 				input.skip();
